feat(frontend): Session management
API to manage a User session, it uses `access` and `refresh` tokens,
the former to authenticate requests and the latter to request a new
updated `access` token. The API has endpoints to manage the:

  * register
  * login
  * refresh of `access` token
  * logout
  * CRUD actions for employees

This is a fully working, albeit demo version of a User Session Management
API. It doesn't handle the refreshing of the `refresh` token, for instance,
every time the user does something, they should have a further 60 minute
session after that, it is only if they do nothing for 60 minutes, that
the session should end. The ability to refresh the `refresh` token will
come next.